Projects/project01/requirements.txt
certifi==2017.11.5
chardet==3.0.4
hlt-client==1.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
zstd==1.3.1
